fix: rebase side effects
rebase 처리로 인해 발생한 문제 정리
- 중복 파일들 제거
- 폴더 구조에 추가된 파일들 편입
