+++ Source/Engine/System/Private/Rendering/Shader/CGraphicShader.cpp
@@ -52,77 +52,25 @@ int CGraphicShader::CreateVertexShader(const wstring& _RelativePath, const strin
 	                           , m_VSBlob->GetBufferSize()
 	                           , nullptr, m_VS.GetAddressOf());
 
-	// InputLayout
-	D3D11_INPUT_ELEMENT_DESC LayoutDesc[8] = {};
+	// 추가 버텍스쉐이더(Instancing 버전) 컴파일
+	hr = D3DCompileFromFile(wstring(SourcePath + _RelativePath).c_str()
+	                        , nullptr, D3D_COMPILE_STANDARD_FILE_INCLUDE
+	                        , (_FuncName + "_Inst").c_str(), "vs_5_0", Flag, 0
+	                        , m_VSInstBlob.GetAddressOf()
+	                        , m_ErrBlob.GetAddressOf());
 
-	LayoutDesc[0].AlignedByteOffset = 0;
-	LayoutDesc[0].Format = DXGI_FORMAT_R32G32B32_FLOAT;
-	LayoutDesc[0].InputSlot = 0;
-	LayoutDesc[0].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[0].InstanceDataStepRate = 0;
-	LayoutDesc[0].SemanticName = "POSITION";
-	LayoutDesc[0].SemanticIndex = 0;
-
-	LayoutDesc[1].AlignedByteOffset = 12;
-	LayoutDesc[1].Format = DXGI_FORMAT_R32G32_FLOAT;
-	LayoutDesc[1].InputSlot = 0;
-	LayoutDesc[1].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[1].InstanceDataStepRate = 0;
-	LayoutDesc[1].SemanticName = "TEXCOORD";
-	LayoutDesc[1].SemanticIndex = 0;
-
-	LayoutDesc[2].AlignedByteOffset = 20;
-	LayoutDesc[2].Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	LayoutDesc[2].InputSlot = 0;
-	LayoutDesc[2].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[2].InstanceDataStepRate = 0;
-	LayoutDesc[2].SemanticName = "COLOR";
-	LayoutDesc[2].SemanticIndex = 0;
-
-	LayoutDesc[3].AlignedByteOffset = 36;
-	LayoutDesc[3].Format = DXGI_FORMAT_R32G32B32_FLOAT;
-	LayoutDesc[3].InputSlot = 0;
-	LayoutDesc[3].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[3].InstanceDataStepRate = 0;
-	LayoutDesc[3].SemanticName = "TANGENT";
-	LayoutDesc[3].SemanticIndex = 0;
-
-	LayoutDesc[4].AlignedByteOffset = 48;
-	LayoutDesc[4].Format = DXGI_FORMAT_R32G32B32_FLOAT;
-	LayoutDesc[4].InputSlot = 0;
-	LayoutDesc[4].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[4].InstanceDataStepRate = 0;
-	LayoutDesc[4].SemanticName = "NORMAL";
-	LayoutDesc[4].SemanticIndex = 0;
-
-	LayoutDesc[5].AlignedByteOffset = 60;
-	LayoutDesc[5].Format = DXGI_FORMAT_R32G32B32_FLOAT;
-	LayoutDesc[5].InputSlot = 0;
-	LayoutDesc[5].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[5].InstanceDataStepRate = 0;
-	LayoutDesc[5].SemanticName = "BINORMAL";
-	LayoutDesc[5].SemanticIndex = 0;
-
-	LayoutDesc[6].AlignedByteOffset = 72;
-	LayoutDesc[6].Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	LayoutDesc[6].InputSlot = 0;
-	LayoutDesc[6].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[6].InstanceDataStepRate = 0;
-	LayoutDesc[6].SemanticName = "BLENDWEIGHT";
-	LayoutDesc[6].SemanticIndex = 0;
-
-	LayoutDesc[7].AlignedByteOffset = 88;
-	LayoutDesc[7].Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	LayoutDesc[7].InputSlot = 0;
-	LayoutDesc[7].InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
-	LayoutDesc[7].InstanceDataStepRate = 0;
-	LayoutDesc[7].SemanticName = "BLENDINDICES";
-	LayoutDesc[7].SemanticIndex = 0;
-
-	if (FAILED(DEVICE->CreateInputLayout(LayoutDesc, 8
-		, m_VSBlob->GetBufferPointer(), m_VSBlob->GetBufferSize()
-		, m_Layout.GetAddressOf())))
+	// 추가 인스턴싱 VS 가 존재한다면
+	if (S_OK == hr)
 	{
+		// 컴파일 된 코드로 Vertex Instancing Shader 객체 만들기
+		DEVICE->CreateVertexShader(m_VSInstBlob->GetBufferPointer()
+		                           , m_VSInstBlob->GetBufferSize()
+		                           , nullptr, m_VSInst.GetAddressOf());
+	}
+
+	if (FAILED(CreateInputLayout()))
+	{
+		assert(nullptr);
 		return E_FAIL;
 	}
